requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

httpx==0.24.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3
